Removed SeriesOID.
